+++ draft-js-emoji-plugin/src/Emoji/index.js
@@ -1,17 +1,18 @@
 import React from 'react';
 import unionClassNames from 'union-class-names';
-import escapeMap from '../utils/escapeMap';
-import mappedUnicode from '../utils/mappedUnicode';
+import emojione from 'emojione';
 
 const Emoji = ({ theme = {}, cacheBustParam, imagePath, className, decoratedText, ...props }) => {
-  const unicode = escapeMap[decoratedText];
-  const backgroundImage = `url(${imagePath}${unicode}.svg${cacheBustParam})`;
+  const shortName = emojione.toShort(decoratedText);
+  // short name to image url code steal from emojione source code
+  const shortNameForImage = emojione.emojioneList[shortName].unicode[emojione.emojioneList[shortName].unicode.length - 1];
+  const backgroundImage = `url(${imagePath}${shortNameForImage}.svg${cacheBustParam})`;
   const combinedClassName = unionClassNames(theme.emoji, className);
   const characterClassName = unionClassNames(theme.emojiCharacter);
   return (
     <span
       className={combinedClassName}
-      title={mappedUnicode[unicode]}
+      title={emojione.toShort(decoratedText)}
       style={{ backgroundImage }}
     >
       <span className={characterClassName}>{props.children}</span>