+++ draft-js-toolbar-plugin/src/decorators/toolbar.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 
 // Get a component's display name
-const getDisplayName = WrappedComponent => {
+const getDisplayName = (WrappedComponent) => {
   const component = WrappedComponent.WrappedComponent || WrappedComponent;
   return component.displayName || component.name || 'Component';
 };
@@ -10,7 +10,7 @@ const getDisplayName = WrappedComponent => {
 let number = 0;
 
 // HoverToolbar decorator will render a toolbar on hovering the WrappedComponent
-export default ({ theme, customRender }) => WrappedComponent => class FocusedToolbarDecorator extends Component {
+export default ({ theme, customRender }) => (WrappedComponent) => class FocusedToolbarDecorator extends Component {
   // Statics
   static displayName = `FocusedToolbar(${getDisplayName(WrappedComponent)})`;
   static WrappedComponent = WrappedComponent.WrappedComponent || WrappedComponent;
@@ -64,7 +64,7 @@ export default ({ theme, customRender }) => WrappedComponent => class FocusedToo
     }
   }
 
-  addActions = actions => {
+  addActions = (actions) => {
     this.componentActions = actions;
   }