+++ draft-js-emoji-plugin/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## To Be Released
+### Added
+
+- Add `emojione` as a dependency.
+- Add `lodash.keys` as a dependency.
+- Update the list of Emojis from Emojione `2.1.2` to `2.2.6`. Thanks to @mzbac
+- The config object will accept a property `priorityList` which should contain Emoji entries used by EmojiOne. These entries will be show first in the EmojiSuggestions dropdown after typing `:`. Thanks to @mzbac
 
 ## 1.2.3 - 2016-07-06